[SPF] Merge configs that are very close
The solver sometimes generate very close configurations (1e-17
difference in coefficients) due to floating point errors when applying
explicit constraints.

However, straight interpolation raises issue when the distance is
calculated to be 0 (due to floating point errors) but two configurations
are not the same. It is safer to check and merge two configurations if
they are extremely close.